modules/module_input/requirements.txt
Pillow
redis
